pyproject.toml
[project]
name = "llm-eval-engine"
version = "0.1.0"
description = "An LLM evaluation engine that scores AI responses using Groq-powered evaluation"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "groq>=1.1.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
